+++ gllm/worker.py
@@ -3,29 +3,17 @@ import torch.distributed as dist
 import traceback
 import logging
 
-from attr import dataclass
 from collections import deque
 from logger import logger
-from transformers.image_utils import load_images
-from gllm.layers.rotary_embedding import MRotaryEmbedding
-from typing import Dict
 
-from gllm.sequence import Sequence
 from gllm.input_data import InputData
 from gllm.model_runner import ModelRunner
 from gllm.comm import zmqComm, IPCPackage
 from gllm.worker_scheduler import WorkerScheduler
-from gllm.dist_utils import (get_local_rank, init_dist, send_pp_data, recv_pp_data, 
+from gllm.dist_utils import (init_dist, send_pp_data, recv_pp_data, 
                              get_rank, get_world_size, is_last_pp_rank,
                              get_pp_size, get_next_pp_rank, get_last_pp_rank,
                              is_output_rank)
-
-@dataclass
-class EmbeddingInfo:
-    embedding: torch.Tensor = None
-    positions: torch.Tensor = None
-    mrope_position_delta:int = None
-    stale: bool = False
 
 # Used with PipeAsyncLLM
 class Worker:
@@ -72,10 +60,6 @@ class Worker:
         self.model_runner.init(self.mp_load_progress)
         self.hidden_size = self.model_runner.model_loader.hidden_size
         self.ret_residual = self.model_runner.model.ret_residual
-        self.use_mm = self.model_runner.use_mm
-        self.model = self.model_runner.model
-        if self.model_runner.use_mm:
-            self.image_processor = self.model_runner.image_processor
         
         if self.rank == 0:
             self.worker_scheduler = WorkerScheduler(
@@ -83,8 +67,6 @@ class Worker:
                 self.model_runner, 
                 self.use_cp_schedule,
             )
-            # embedding cache: seq_id => embedding
-            self.embedding_cache:Dict[int,EmbeddingInfo] = {}
         elif self.pp_rank == 0:
             # Input data for each rank except 0
             self.schedule_queue = deque()
@@ -100,14 +82,11 @@ class Worker:
 
     # driver worker => other workers
     def recv_schedule_seqs(self):
-        seqs_data = self.comm.recv_schedule_seqs()
-        if seqs_data is not None:
-            seqs, positions = seqs_data
-            if positions is not None:
-                positions = positions.to(f'cuda:{get_local_rank()}')
+        seqs = self.comm.recv_schedule_seqs()
+        if seqs is not None:
             self.schedule_queue.append(
                 InputData(seqs, self.model_runner.memory_manager,
-                          use_mla=self.use_mla, positions=positions))
+                          use_mla=self.use_mla))
 
     # pp last rank => pp next rank
     def recv_intermediate_data(self):
@@ -172,83 +151,18 @@ class Worker:
             output = self.model_runner.step_once(input_data)
             if get_pp_size() != 1:
                 send_pp_data(output, get_next_pp_rank())
-                
-    
-    @torch.inference_mode()
-    def mm_prepare_inputs(self, seqs: list[Sequence]):
-        # Calculate the embedding and positions of pic
-        batch_embeddings = []
-        batch_positions = []
-        for seq in seqs:
-            embedding = None
-            position = None
-            if seq.computed_prompt:
-                embedding_info = self.embedding_cache[seq.seq_id]
-                assert embedding_info.stale
-                position = MRotaryEmbedding.get_next_input_positions(
-                    embedding_info.mrope_position_delta,
-                    seq.computed_token_num,
-                    seq.seq_len
-                )
-                position = torch.tensor(position)
-                embedding = self.model.get_input_embeddings(
-                    torch.tensor(seq.token_ids[seq.computed_token_num:seq.seq_len]))
-            else:
-                embedding_info = None
-                if seq.seq_id not in self.embedding_cache or self.embedding_cache[seq.seq_id].stale:
-                    mm_embeddings = None
-                    image_grid_thw = None
-                    if seq.mm_contents is not None:
-                        images = load_images(seq.mm_contents)
-                        images_input = self.image_processor(images=images)
-                        image_grid_thw = images_input['image_grid_thw']
-                        mm_embeddings = self.model.get_multimodal_embeddings(**images_input)
-                    prompt_embeddings = self.model.get_input_embeddings(
-                        torch.tensor(seq.token_ids), mm_embeddings)
-                    prompt_positions, mrope_position_delta = MRotaryEmbedding.get_input_positions(
-                        input_tokens=seq.token_ids,
-                        hf_config=self.model.config,
-                        image_grid_thw=image_grid_thw,
-                        video_grid_thw=None,
-                        second_per_grid_ts=None,
-                    )
-                    embedding_info = EmbeddingInfo(
-                        prompt_embeddings, prompt_positions, mrope_position_delta)
-                    self.embedding_cache[seq.seq_id] = embedding_info
-                    position = prompt_positions[
-                        :, seq.computed_token_num:seq.seq_len]
-                    embedding = prompt_embeddings[
-                        seq.computed_token_num:seq.seq_len, :]
-                else:
-                    embedding_info = self.embedding_cache[seq.seq_id]
-                    position = embedding_info.positions[
-                        :, seq.computed_token_num:seq.seq_len]
-                    embedding = embedding_info.embedding[
-                        seq.computed_token_num:seq.seq_len, :]
-                if seq.seq_len == seq.prompt_len:
-                    # invalidate embedding_cache
-                    embedding_info.stale = True
-                    embedding_info.embedding = None
-                    embedding_info.positions = None
-            batch_embeddings.append(embedding)
-            batch_positions.append(position)
-        input_embeddings = torch.concat(batch_embeddings)
-        positions = torch.concat(batch_positions, dim=1)
-        return input_embeddings, positions
     
     def schedule_forward(self):
         schedule_seqs = self.worker_scheduler.schedule_once()
         if len(schedule_seqs) != 0:
             input_embeddings = None
             positions = None
-            if self.use_mm:
-                input_embeddings, positions = self.mm_prepare_inputs(schedule_seqs)
             input_data = InputData(
                 schedule_seqs, self.model_runner.memory_manager,
-                use_mla=self.use_mla, positions=positions)
+                use_mla=self.use_mla)
             if get_world_size() > 1:
-                self.comm.send_schedule_seqs((schedule_seqs, positions))
-            output = self.model_runner.step_once(input_data, input_embeddings=input_embeddings)
+                self.comm.send_schedule_seqs(schedule_seqs)
+            output = self.model_runner.step_once(input_data)
 
             if not is_output_rank():
                 send_pp_data(output, get_next_pp_rank())
